mpi/info: split MPIR implementations from MPI functions
