+++ src/mpi/info/info_set.c
@@ -122,49 +122,3 @@ int MPI_Info_set(MPI_Info info, const char *key, const char *value)
     goto fn_exit;
     /* --END ERROR HANDLING-- */
 }
-
-
-#ifndef MPICH_MPI_FROM_PMPI
-
-int MPIR_Info_set_impl(MPIR_Info * info_ptr, const char *key, const char *value)
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPIR_Info *curr_ptr, *prev_ptr;
-    MPIR_FUNC_TERSE_STATE_DECL(MPID_STATE_MPIR_INFO_SET_IMPL);
-
-    MPIR_FUNC_TERSE_ENTER(MPID_STATE_MPIR_INFO_SET_IMPL);
-
-    prev_ptr = info_ptr;
-    curr_ptr = info_ptr->next;
-
-    while (curr_ptr) {
-        if (!strncmp(curr_ptr->key, key, MPI_MAX_INFO_KEY)) {
-            /* Key already present; replace value */
-            MPL_free(curr_ptr->value);
-            curr_ptr->value = MPL_strdup(value);
-            break;
-        }
-        prev_ptr = curr_ptr;
-        curr_ptr = curr_ptr->next;
-    }
-
-    if (!curr_ptr) {
-        /* Key not present, insert value */
-        mpi_errno = MPIR_Info_alloc(&curr_ptr);
-        MPIR_ERR_CHECK(mpi_errno);
-
-        /*printf("Inserting new elm %x at %x\n", curr_ptr->id, prev_ptr->id); */
-        prev_ptr->next = curr_ptr;
-        curr_ptr->key = MPL_strdup(key);
-        curr_ptr->value = MPL_strdup(value);
-    }
-
-  fn_exit:
-    MPIR_FUNC_TERSE_EXIT(MPID_STATE_MPIR_INFO_SET_IMPL);
-    return mpi_errno;
-
-  fn_fail:
-    goto fn_exit;
-}
-
-#endif /* MPICH_MPI_FROM_PMPI */
